requirements.txt
discord.py-Console
openai
discord
google-api-python-client
mysql-connector